Phase 4.1 — Complete hive_session_t struct Implements Task 4.1 per DEVELOPMENT.md and ARCHITECTURE.md §2. - Completely rewrites hive_internal.h with all 10 struct regions (A–J):   A: identity/role/state (session_id, role, session_state)   B: peer settings (peer_settings hive_settings_t)   C: local settings (local_settings, local_settings_pending, settings_pending_ack)   D: flow control (send_window, recv_window, recv_consumed)   E: HPACK tables (enc_table/dec_table hpack_entry_t arrays + sizes/caps)   F: stream table (streams array, hash table, active/reserved counts)   G: send queue (send_buf, send_iov, send_buf_used/cap/count, send_partial/offset)   H: recv state machine (recv_buf, recv_used, recv_frame_hdr, continuation_sid)   I: callbacks (on_headers, on_data, on_send, on_error, user_data)   J: connection error (goaway_error_code) - Adds stream_hash_entry_t (8 bytes): two-layer hash map entry   with sentinel constants STREAM_HASH_EMPTY (0xFFFFFFFF) and   STREAM_HASH_TOMBSTONE (0xFFFFFFFE) - Adds hive_stream_state_t enum (STREAM_IDLE through STREAM_CLOSED) - Adds hive_stream_t with precise 64-byte layout enforced by _Static_assert:   stream_id(4) + state(1) + flags(1) + weight(1) + _pad(1) +   send_window(4) + recv_window(4) + recv_consumed(4) + _pad2(4) +   content_length_expected(8) + content_length_received(8) +   user_data(8) + data_source(16) = 64 bytes - Adds four _Static_assert checks:   sizeof(hive_stream_t) == 64   sizeof(stream_hash_entry_t) == 8   sizeof(hpack_entry_t) == 8   sizeof(hive_settings_t) == 24 - Changes send_iov_count from uint32_t to int per ARCHITECTURE.md §2.10;   updates test assertions accordingly - Fixes hive_frame.c narrowing warning: s->role = (uint8_t)role Tests: 91/91 total (no new tests required; struct layout verified by _Static_assert) Docs: - Mark Task 4.1 as DONE in DEVELOPMENT.md - Advance Next Task to 4.2 Linux/OpenBSD: make clean + make dev + make test pass (91/91); Linux make valgrind + make test-tsan + make lint + make format clean.

+++ src/hive_frame.c
@@ -226,7 +226,7 @@ void
 frame_recv_init(hive_session_t *s, hive_role_t role)
 {
 	zero_bytes((uint8_t *)s, sizeof(*s));
-	s->role = role;
+	s->role = (uint8_t)role;
 	s->opt_max_frame_size = 16384;
 	s->opt_max_continuation_size = 65536;
 	s->local_settings.max_frame_size = 16384;
